refactor: reorganize and optimize code related to page layout components refactor: reorganize exports in auth-page, application-layout, and authentication-provider-button modules
