+++ packages/ui/src/components/application-layout/index.ts
@@ -1,0 +1,1 @@
+export * from './page'
