+++ packages/ui/src/components/auth-page/index.ts
@@ -1,4 +1,4 @@
-import { FooterFlexLine, PageFooterWrapper } from "./footer";
+import { FooterFlexLine, PageFooterWrapper } from "./page-footer";
 import { FormFooter, FormWrapper } from "./input-wrapper";
 import { FormHeader, PageWrapper } from "./page-wrapper";
