+++ packages/ui/src/components/authentication-provider-button/index.ts
@@ -1,0 +1,1 @@
+export * from './button'
