+++ packages/ui/src/components/index.ts
@@ -1,3 +1,4 @@
-export * from "./auth-page/auth-page";
-export * from "./application-layout/application-page";
-export * from "./authentication-provider-button/auth-provider-button";
+export * from "./auth-page";
+export * from "./application-layout";
+export * from "./authentication-provider-button";
+
